move getTokenAttributes
