+++ libs/evm-protocols/src/index.ts
@@ -2,5 +2,6 @@ export * from './abis';
 export * as commonProtocol from './common-protocol';
 export * from './common-protocol/contractHelpers/namespace';
 export * from './common-protocol/contractHelpers/stake';
+export * from './common-protocol/contractHelpers/tokens';
 export * from './event-registry/eventRegistry';
 export * from './event-registry/eventSignatures';
