+++ libs/model/test/contest/contests-projection-lifecycle.spec.ts
@@ -6,7 +6,7 @@ import {
   handleEvent,
   query,
 } from '@hicommonwealth/core';
-import { commonProtocol } from '@hicommonwealth/evm-protocols';
+import * as evm from '@hicommonwealth/evm-protocols';
 import { createEventRegistryChainNodes, models } from '@hicommonwealth/model';
 import { ContestResults, EventNames } from '@hicommonwealth/schemas';
 import { delay } from '@hicommonwealth/shared';
@@ -24,13 +24,12 @@ import {
 import { z } from 'zod';
 import { Contests } from '../../src/contest/Contests.projection';
 import { GetAllContests } from '../../src/contest/GetAllContests.query';
-import {
-  contestHelper,
-  contractHelpers,
-} from '../../src/services/commonProtocol';
+import { contestHelper } from '../../src/services/commonProtocol';
 import { seed } from '../../src/tester';
 
 chai.use(chaiAsPromised);
+
+const { commonProtocol } = evm;
 
 // TODO: re-enable test
 describe('Contests projection lifecycle', () => {
@@ -65,7 +64,7 @@ describe('Contests projection lifecycle', () => {
   const decimals = commonProtocol.WeiDecimals[commonProtocol.Denominations.ETH];
   const topic_id = 100;
 
-  const getTokenAttributes = vi.spyOn(contractHelpers, 'getTokenAttributes');
+  const getTokenAttributes = vi.spyOn(evm, 'getTokenAttributes');
   const getContestScore = vi.spyOn(contestHelper, 'getContestScore');
   const getContestStatus = vi.spyOn(contestHelper, 'getContestStatus');
